add page next & mr_result
